Fix flickering with data loader (#348)
* added loader

* added loader

* linter

* fix name
